# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "HospitalSupplyCategory": {
        "type": "string",
        "enum": [
          "medications",
          "iv_fluids",
          "medical_supplies",
          "soft_foods",
          "water",
          "beds_capacity",
          "lab_diagnostics",
          "transport",
          "other"
        ]
      },
...
      "HospitalSupplyUrgencySemaphore": {
        "type": "string",
        "enum": [
          "yellow",
          "red",
          "unknown"
        ],
        "description": "Urgencia permitida para necesidades activas; verde se reserva para semáforos sin necesidad crítica."
      },
...
      "HospitalSupplyNeed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/HospitalSupplyCategory"
          },
          "categoryLabel": {
            "type": "string"
          },
          "itemType": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "nullable": true
          },
          "unit": {
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/HospitalSupplyUrgencySemaphore"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "partially_covered",
              "covered",
              "cancelled",
              "needs_verification"
            ]
          },
          "publicNote": {
            "type": "string"
          },
          "lastConfirmedAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          },
          "updatedAgo": {
            "type": "string"
          }
        }
      },
      "HospitalSupplyNeedRestricted": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HospitalSupplyNeed"
          },
          {
            "type": "object",
            "properties": {
              "restrictedNote": {
                "type": "string"
              },
              "updatedBy": {
                "type": "string"
              },
              "source": {
                "type": "string"
              },
              "createdAt": {
                "type": "integer"
              }
            }
          }
        ]
      },
...
      "HospitalSupplyNeedPatchInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "partially_covered",
              "covered",
              "cancelled",
              "needs_verification"
            ]
          },
          "publicNote": {
            "type": "string"
          },
          "restrictedNote": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
      "HospitalSupplyHelpRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/HospitalSupplyCategory"
          },
          "categoryLabel": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/HospitalSupplyUrgencySemaphore"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "contacting",
              "resolved",
              "closed"
            ]
          },
          "requestedBy": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "restrictedNote": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          },
          "updatedAgo": {
            "type": "string"
          }
        }
      },
      "HospitalSupplyHelpRequestInput": {
        "type": "object",
        "required": [
          "category",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/HospitalSupplyCategory"
          },
          "message": {
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/HospitalSupplyUrgencySemaphore"
          },
          "requestedBy": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "restrictedNote": {
            "type": "string"
          }
        }
      },
      "HospitalSupplyHelpPatchInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "contacting",
              "resolved",
              "closed"
            ]
          },
          "restrictedNote": {
            "type": "string"
          },
          "requestedBy": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/hospitals/{id}/supplies/help/{requestId}": {
      "patch": {
        "tags": [
          "hospitals"
        ],
        "summary": "Actualiza una solicitud restringida de ayuda de insumos (requiere x-admin-token o x-hospital-poc-token del hospital)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HospitalSupplyHelpPatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request": {
                      "$ref": "#/components/schemas/HospitalSupplyHelpRequest"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (falta token admin o POC activo para el hospital)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Hospital o solicitud no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload demasiado grande",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo actualizar la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/hospitals/{id}/supplies/help": {
      "post": {
        "tags": [
          "hospitals"
        ],
        "summary": "Crea una solicitud restringida de ayuda para actualizar insumos (requiere x-admin-token o x-hospital-poc-token del hospital)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HospitalSupplyHelpRequestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitud creada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request": {
                      "$ref": "#/components/schemas/HospitalSupplyHelpRequest"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (falta token admin o POC activo para el hospital)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Hospital no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload demasiado grande",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo guardar la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/hospitals/{id}/supplies/needs/{needId}": {
      "patch": {
        "tags": [
          "hospitals"
        ],
        "summary": "Actualiza el estado operativo de una necesidad específica (requiere x-admin-token o x-hospital-poc-token del hospital)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "needId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HospitalSupplyNeedPatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Necesidad actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "need": {
                      "$ref": "#/components/schemas/HospitalSupplyNeedRestricted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (falta token admin o POC activo para el hospital)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Hospital o necesidad no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload demasiado grande",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo actualizar la necesidad",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}